Componentのレイヤーへの投稿
